% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{caretEnsNestedCV}
\alias{caretEnsNestedCV}
\title{Nested CV with ensemble modeling.}
\usage{
caretEnsNestedCV(ml.out, cv.folds, ds)
}
\value{
Performance values.
}
\description{
\code{caretEnsNestedCV} Prototype ensemble modeling with nested CV>
}

