% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModels.R
\name{compareModels}
\alias{compareModels}
\title{Compare evaluation metrics for cross-validation on the training set and on the test set for a caret model. Return the result as a ggplot object.}
\usage{
compareModels(
  model_list,
  test_data = NULL,
  model_order = "RMSE",
  title = "Comparing model metrics"
)
}
\arguments{
\item{model_list}{A named list of caret models.}

\item{test_data}{The test set of data for evaluation.}

\item{model_order}{an evaluation metric as a string, either "RMSE", "MAE" or "Rsquared". Default "RMSE".}

\item{title}{A string to title the resulting ggplot graph.}
}
\value{
a ggplot object.
}
\description{
Compare evaluation metrics for cross-validation on the training set and on the test set for a caret model. Return the result as a ggplot object.
}
\examples{
compareModels(model_list = allModels,
              test_data = ames_test_engineered,
              model_order = "RMSE")
}
