% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareModelsDf.R
\name{compareModelsDf}
\alias{compareModelsDf}
\title{Compare evaluation metrics for cross-validation on the training set and on the test set for a caret model, return the result as a tibble.}
\usage{
compareModelsDf(model_list, test_data = NULL, model_order = "RMSE")
}
\arguments{
\item{model_list}{A named list of caret models.}

\item{test_data}{The test set of data for evaluation.}

\item{model_order}{an evaluation metric as a string, either "RMSE", "MAE" or "Rsquared". Default "RMSE".}
}
\value{
a tibble with all the metrics for each model on the training and test data.
}
\description{
Compare evaluation metrics for cross-validation on the training set and on the test set for a caret model, return the result as a tibble.
}
\examples{
compareModelsDf(model_list = allModels,
              test_data = ames_test_engineered,
              model_order = "RMSE")
}
