% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{covarConvert}
\alias{covarConvert}
\title{Convert covariates for machine learning.}
\usage{
covarConvert(clin.covars)
}
\arguments{
\item{clin.covars}{input data.frame}
}
\value{
model matrix for learning.
}
\description{
\code{covarConvert} takes none numeric and categories and converts them all for learning.
}

