% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{extractCVres}
\alias{extractCVres}
\title{Extract out of sample CV predictions.}
\usage{
extractCVres(mod, average = TRUE)
}
\arguments{
\item{mod}{a caret model}

\item{average}{TRUE or FALSE; averages over repeats if TRUE}
}
\value{
data frame of predictions
}
\description{
\code{extractCVres} This extracts the out of sample cross validated predictions from the caret model object.
}

