% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{nestedCaretCVPerf}
\alias{nestedCaretCVPerf}
\title{Nested CV performance.}
\usage{
nestedCaretCVPerf(ml.out, cv.folds)
}
\arguments{
\item{ml.out}{models from caret.}
}
\value{
Performance values.
}
\description{
\code{nestedCaretCVPerf} Prototype for extracting performance from nested CV.
}

