% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{plotVIP}
\alias{plotVIP}
\title{Plot VIP values.}
\usage{
plotVIP(model, nFeats.plot)
}
\arguments{
\item{model}{to use}

\item{nFeats.plot}{number of features to plot}
}
\value{
a barplot
}
\description{
\code{plotVIP} Make a plot for the top VIP.
}

