% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/funcs.r
\name{processCVres}
\alias{processCVres}
\title{Average predictions from repeats.}
\usage{
processCVres(cv.res)
}
\arguments{
\item{cv.res}{the output of extractCVres.}
}
\value{
A matrix of cross validated, averaged, predicted values.
}
\description{
\code{processCVres} averages the predictions from multiple repeats together.  Should generally just be called using average param in extractCVres.
}

